Altered contract function to use alias method for efficient sampling from a discrete distribution
- Created class SimpleTuple
	- Simple 3-tuple data structure for (prob, x, y)
	- Used in alias method

- Vertex class
	- Added partition method and partitioned boolean to check if Vertex is partitioned
		- Partitions edge probabilities for efficient sampling
	- Added partitioned boolean
		- Used to check if Vertex is partitioned
		- Set to false initially and each time an edge is added or removed
		- Indicates that Vertex edges need repartitioning
	- Added method sample that samples from edge partitions

- Graph class
	- Sparsify method is now "generic" and can use either contraction method, Gaussian or random edge contraction
	- Added preprocessing step that partitions edge probabilities
	- Rewrote contract method as it was quite messy

- Updated timelog

+++ GraphCompression/src/graphComponents/AdjNode.java
@@ -4,9 +4,9 @@ package graphComponents;
 /*Represents an adjacent Vertex in the adjacency list of a Vertex*/
 public class AdjNode {
 	private Vertex v;	//Index of the Vertex represented by this node in the Graph's Vertex list
-	private float weight;	//Represents the weight of the edge to get from some Vertex to this adjacent Vertex
+	private double weight;	//Represents the weight of the edge to get from some Vertex to this adjacent Vertex
 	
-	public AdjNode(Vertex v, float weight) {
+	public AdjNode(Vertex v, double weight) {
 		this.v  = v;
 		this.weight = weight;
 	}
@@ -25,13 +25,13 @@ public class AdjNode {
 	
 	
 	/*Returns weight of the edge to reach this Vertex from some adjacent Vertex*/
-	public float getWeight() {
+	public double getWeight() {
 		return this.weight;
 	}
 	
 	
 	/*Updates the value of weight*/
-	public void setWeight(float newWeight) {
+	public void setWeight(double newWeight) {
 		this.weight = newWeight;
 	}
 }
